Create a list of follow up messages
Since not everyone responds immediately, for lack of time or because of an interrupt, it is sometimes necessary to send reminders to bump ourselves back to the top of the message queue. This is tricky, because a real person is supposed to be sending the messages, and repeating the same one will be seen as the act of a bot. If we don't do this, we'll be forgotten and never in contact.

I suggesting we use this series of messages that build upon the previous with the assumption that the previous one was read but not answered. 

We'll want to add some abstraction so the messages can be used outside of San Francisco.