Add browser authentification
Previously, users had to use my Facebook Token and ID to authenticate to
Tinder. This was a problem since my credentials were exposed.

* Prompt user for Facebook login and password
* Add Watir webdriver to authenticate to Facebook
* Retrieve ID and Token

This still needs to be refined. For instance the webdriver only works
with Firefox so far. Also, when typing the Facebook password, it should
be hidden. Last but not least, we need to find a way to protect those
credentials.